API-1520: Extract download logs button to separate component
